requirements.txt
yfinance
pandas
pandas_ta
python-telegram-bot
requests